Initial pbench-agent container image layering
[Back-port]

There are 3 layers: a "base" layer, intermediate layers on top of base
providing "tools" and "workloads" independently, and an "all" layer that
pulls them all together.

The "base" container image is a full install of the pbench-agent RPM
without and RPMs for tools or workloads, just the core components of the
pbench-agent.

The "tools" and "workloads" container images contain RPMs required for
the operation of the various tools and workloads respectively.

The "all" image is a convenient combination of the "tools" and
"workloads" images into one image.

Having the tools and workloads dependencies in separate images allows
users to construct their own images when they have specific requirements
without pulling in extra unwanted dependencies.

The images will be tagged using the following image tags automatically
when built:

 * `<git commit sha1>`
 * `<rpm version #>`

A separate manual process to label the set of images as the `latest` is
offered with the `make` target `latest`.  We also `make` targets for
labeling the latest "major" and "major.minor" versions as
`v<major>-latest` and `v<major>.<minor>-latest` (respectively) with the
`major-latest` and `major-minor-latest` targets.

Future work is needed in order to reduce the size of the "base" image.