Prepare roles for Ansible Galaxy
We have a reserved "pbench" namespace on Galaxy, so this is the
"agent" collection under the "pbench" namespace. Galaxy uses
"pbench.agent" most of the time, but there are cases where you have to
tell it "pbench/agent" in order for it to understand what you are
talking about.

- Rename all the roles using underscores instead of dashes.

- Rename the playbooks as well.

- Change the directory structure according to what `ansible-galaxy
  collection init pbench.agent' gives.

- Add galaxy info to all the meta/main.yml files (make sure
  indentation is correct - DAMHIKT).

- Add README.md files to all the roles.

- Add galaxy.yml file to the collection's top dir.

- Delete remote_user from all the playbooks.

There is a simple Makefile that builds a tarball, runs the sanity
tests and uploads the tarball to Galaxy. The sanity test
("ansible-test sanity") is absolutely indispensable: Galaxy is very
picky about what it will allow - in particular indentation errors are
fatal and cause cryptic error messages and trailing empty lines are
also fatal but are diagnosed more precisely. This imposes a "yamllint"
dependency, but on the plus side I often invoked that by hand to avoid
the "bad telephone line" effect of having Galaxy (or even
ansible-test) tell me what's wrong.

Published the current version as 1.0.3.