Use click context
Whenever a click command is executed a context object is created.
This context holds the state for the particular command that is
executed. It remembers passed parameters, which command created it,
and what resources that are needed to be cleared up.

We use the click context so we can manage the CLI a bit better
and use it later to manage the unified CLI.

Signed-off-by: Charles Short <chucks@redhat.com>