Minimal attempt to remove tool meister commands
Here we replace the `pbench-tool-data-sink` command line interface entirely
replacing it with a `fork()` and invocation of the Tool Data Sink library
code directly.  We then do the same for the `pbench-tool-meister` command.

In both cases the original commands are copied as-is to the library side of
the tree with no modifications.  The new `pbench-tool-meister` command is a
very simple trampoline command file required for our SSH tool meister
default orchestration.

This demonstrates that neither the Tool Data Sink or the Tool Meister
require a CLI interface for their operation as they only need a simple set
of arguments passed to them.