Fix tool argument handling
Fixes issue #1751, where tools with multiple arguments only have the
first argument passed to the tool command.

When we load the registered tools into Redis, we process the tool
options to make sure they are handled as a single string on the tool
command line.

We alos have to ensure tool option order is preserved.  When
`pbench-register-tool` accepts command line arguments to add as tool
options, the order of those options received needs to be preserved.

The code needs to use indexed arrays properly to preserve order.  It was
trying to use associative arrays as indexed.