Ensure tool option order is preserved
When `pbench-register-tool` accepts command line arguments to add as
tool options, the order of those options received needs to be preserved.

The code needs to use indexed arrays properly to preserve order.  It was
trying to use associative arrays as indexed.

We also add a unit test for `pbench-start-tools`, using the `tcpdump`
command with two arguments registered.