Fix bench/util-scripts unit tests env overrides
Both the `bench-scripts` and the `util-scripts` unit test environments
were overriding the various `pbench_*` environment variables defined in
`base` with definitions provided ahead of time.  While this works, and
the `base` script allows for it, it hid a few minor problems with the
way some tests used sample data vs config data.

By adding an explicity definition of `pbench_run` in the generated
`pbench-agent.cfg` files, we avoid the need to override many of the
`pbench_*` environment variables.