Setup base classes
Stub in base classes for various CLI utils that we will
be replacing shell scripts with python scripts.

Since there are alot of different shell scripts to convert
I grouped the various commands into sub-classes from the BaseCommand
since alot of the shell scripts are re-using the same code as well.
So it made sense to me.

Signed-off-by: Charles Short <chucks@redhat.com>