Refactor agent RPM infrastructure to DRY
We take a page from the `b0.69` branch, and refactor the Makefile to avoid
having to build the source tree twice, once for the source tar ball, once
when running the RPM build.  Now we build the install tree for the source
tar ball and just use that directly.

We also refactor the agent source tree `Makefile` so that we only add what
we need to the source tar ball, allowing us to drop the use of the
`exclude.pats` file.

Further, we switch to using a Jinja2 template for the `pbench-agent.spec`
file, allowing us to drop the need for the `set-version-release` script.