Tell pbr to stop creating AUTHORS & ChangeLog
The `pbr` sub-system from OpenStack auto generates an `AUTHORS` file
and a `ChangeLog` file under the assumption that any time `pbr` is
invoked one would want them.  Unfortunately, we don't want them when
we run our tests.

While we are in the `tox.ini` we remove the individual `usedevelop`
settings since the "global" `testenv` settings are shared by all.