Ensure proper use of rh-python36 only for RHEL 7
Change the RPM `.spec` file to uncomment the `scl_source enable` line
in the `/opt/pbench-agent/base` shell script for RHEL 7 environments,
and remove all other logic to detect a Python3 interpreter.  By doing
this we can create one RPM for all RHEL 7 systems and still have the
`base` script work in the face of a natively installed Python3
interpreter (RHEL 7.7 and later) along side an SCL rh-python36
installation, while still supported Fedora, CentOS, and RHEL 7.6 (and
earlier) installations.

We also add direct dependency on *both* the pip RPMs and the Python 3
RPMs, since the Python3 requirement placed through the pip RPMs doesn't
seem end up recording the `pbench-agent` has a requirement for Python 3.
That is to say, one can uninstall Python3 and the `pbench-agent` RPM
won't be flagged as a dependency.

Background on RHEL 7, SCL, and Python 3:

In most environments where the `pbench-agent` RPM is installed, that
includes even the `pbench-agent` development environment, we expect that
the `python3` command is available, and that the associated `pip3`
command will install Python3 packages into a `site-packages` location on
the default `PYTHONPATH` used by that interpreter.

However, for RHEL 7 we have a complicated situation.  Prior to RHEL 7.7,
the only way to get a native Python3 interpreter was to either build it
yourself, or install it from EPEL.  This was not desirable from a
supportability standpoint because some environments explicitly disallow
the use of EPEL, and won't build, install, and maintain their own copy
of a Python3 interpreter.  Since the Software Collections Library
started to provide a supported Python 3.6 interpreter, we used that
environment when it became available.

When RHEL 7.7 was released [1], a native Python3 installation was made
available, one that did not have to be installed vi the SCL.

This might have been a nice solution to the problem of having to do
something different for RHEL 7 environments, but since we have to
support versions of RHEL before 7.7, we still need to use the SCL
because the `Requires` of an RPM `.spec` file is not an "or" operation:
`.spec` files have no way to express "we depend on _either_ rh-python36
or python3" RPMs.  Since SCL is available across all releases of RHEL
7, we must use the SCL.

A further complication comes from the fact that the SCL is available
on Fedora, CentOS, and RHEL platforms where an installation can have
*both* a native Python 3 interpreter and one provided by a
`rh-python36` SCL installed collection.  The previous code would first
look for a `python3` command in the `PATH`, and if found, blindly use
that.  But that meant that on a RHEL 7 box, where the RPM would have
previously installed all the required modules in the `rh-python36`
environment, a user would encounter a slew of "Module not found"
errors when the Python 3 interpreter would start to execute
`pbench-agent` code.

The previous simplistic logic to look for a `python3` command first,
then, if not found, look for an SCL environment, would not work when
BOTH of those environments were present.  Regardless of which one
was checked for first, depending on the platform installation, we'd
still get it wrong.

Since a rh-python36 SCL environment provides a `python3` command as
well, we remove the SCL detection logic entirely from `base`, leaving
behind a commented out `source scl_source enable rh-python36` command.
We then modify the `pbench-agent.spec` file to explicitly uncomment
that line in the RPM post-install phase (`%post`) on RHEL 7 installs.

[1] https://access.redhat.com/documentation/en-us/red_hat_enterprise_linux/7/html-single/7.7_release_notes/index#enhancement_compiler-and-tools