Use the %config for the agent config file
Instead of using `%ghost`, which means [1] the file is not provided by
the RPM package, but when created is owned by the RPM, and hence removed
when the RPM is removed, we use `%config(noreplace)`, which is dedicated
for use with configuration files.  If the user does not modify the file,
the RPM will remove it when uninstalled, and will update it with a new
one when the RPM is updated. However, if the user makes its own modifi-
cations to the config file, the RPM packaging will not remove it on
uninstall, placing the updated copy in `*.rpmnew`.

[1] http://ftp.rpm.org/max-rpm/s1-rpm-inside-files-list-directives.html