Move md5sum() method to common utils
The `md5sum()` method was hosted in the "server" tree of the Python
modules, but "agent" code was also referencing it.  We have moved it
to a new `common` path allowing both the agent and the server to
reference it.  The packaging steps just ignore the agent and server
trees, as appropriate, so the new common tree will be present for
both.