Add a "common" unit test environment
With the `md5sum()` method being shared between the agent and the
server, we need a test that is independent of either.

By adding a test for the `md5sum()` method, we realized that we
should be testing this new "common" test under both the agent and
the server environments.  So we have modified the tox environments
so that both `py3-agent` and `py3-server` test the common module,
and test their own modules.

While looking through the tests, the functional tests also need
to be run under both the agent and server environments. We drop the
separate `py3-functional` tox environment, and just run it with
both the `py3-agent` and `py3-server` environments.

We also remove the unused `unit/server/common.py` file.