Remove use of `screen` from Tool Meister
Given that `screen` has grown a little "long in the tooth", and that for RHEL 7
the version of `screen` is so old that it does not contain a fix for a race
condition on its socket directory creation, it seems the best course of action
is to remove the use of it from the Tool Meister code.

This might mean that certain tools that require a pseudo terminal will stop
working, but solving that problem instead of fixing `screen` seems to be more
approachable.