Add support for sysinfo via Tool Meister
The `pbench-collect-sysinfo` CLI interface has been modified to send a
message to the Tool Meisters to have them collect all the requested
system configuration information.  Much like tool data is collected and
sent back to the Tool Data Sink from each Tool Meister, the system
information is collected by each Tool Meister and sent back to the Tool
Data sink to be stored in the requested location.

This means that `pbench-collect-sysinfo` no longer uses `ssh` to collect
and copy the system information from remote hosts.  As it is for tools
running on the local host, the local Tool Meister write the collected
data into the local volume directly without sending it through the Tool
Data Sink.

This change allows us to delete the `pbench-remote-sysinfo-dump`
trampoline interface since the Tool Meister takes are of remote
operations.

Each benchmark-script has been modified to invoke the
`pbench-collect-sysinfo` following the `pbench-tool-meister-start` call,
and before the `pbench-tool-meister-stop` call.

We fixed a small bug for `pbench-trafficgen` which was overwriting the
"beginning" (`.../sysinfo/beg`) collected system information when
collecting the "ending" (`.../sysinfo/end`) information.