Backport container build enhancements
There have been a number of enhancements to how we build containers
that need to be available for building the `b0.69` branch.  We were
successful using these changes to build the `v0.69.4-agent` release
using them.

We had to remove the `atomic-openshift-clients` RPM dependency from
the "tool" image since that repo is really per OpenShift version.