Remove SCL from the agent side
(Backport of PR #1915 & #1916)

Remove all references to "SCL" throughout the code, relying on packaging
and an installation procedures to provide a proper Python 3 environment.

We use Python 3.6 explicitly for RHEL 8 / CentOS 8 to avoid packages
conflicts which don't work with one or the other (e.g. python3-requests
is an RPM that only works with Python 3.6, not Python 3.8).

We have refined the RPM list as a result for this work, and dropped the
code to generate different "final" `requirements.txt` files based on the
platform.  This is because a `pip3 install` will only install modules
that are not already found on the `PYTHONPATH`.  RPMs which satisfy the
module requirements will prevent `pip3` from installing it locally.

Further, dnf caches repo states in weird ways, so when building "all" the
images, we use a new `all-tags` target which helps ensure we get the same
tags for all images.  We want individual tags per image in general, since
we don't want to force somebody to have to build everything.  But *when*
building everything we need the consistency.

Finally, we add the `all-dockerfiles` convenience target to build all
the `.repo` and `.Dockerfile` files that would be used.