Reformat the HTML to separate content from markup.
This commit is a reformat-only commit of start.html. The changes are:

- Reduce indentation so that long lines of markup (mostly) don't wrap
  on a decent-sized monitor.

- Separate markup from content, particularly in <p>...</p> sections
  which we reformat like this:

  <p>
  ...Content...
  </p>

- Use fill-paragraph in emacs to break the content of each paragraph
  into short lines that don't wrap and, most importantly, do not
  contain markup. Two exceptions: there is emphasis using
  <spann>...</span> and links using <a ...>...</a> within some of the
  content.

- Finally, we clean up white space.

The result is that content is clearly separated from markup, at least
for the descriptive text paragraphs. Section titles and such are still
lost in the thicket of markup, but it should be easier to find them and
fix them.

Finally, all this reformatting is a no-op as far as rendering goes: the
resulting rendered pages are before-and-after identical.