Address review comments
Addresses most of  @portante's comments on the content, with the
exception of the comments on the "Best Practices" tab, which have
been summarized in issue #1927.

Addresses @anishaswain's comments on the structure in the sense that
all the tabs show up and have the correct content. But I'm a noob
on patternfly and there are probably some major goofs in the structure
(but they don't prevent the content from showing more or less correctly).