addition of function description
The tab names,are not unique by the implementation and are given unique id to each of them according to their names(i.e id of setup is "setup", id of "best practices is "best").

In order to add a new tab, we just need to make sure that the `content_div` id( i.e the content division needed to be displayed) is unique and the tabLink (i.e the id of each tab button) is the combination of the `content_div` id and "Link"

for example, in the setup tab (line no. 67 in start.html);
`<button id="setupLink" class="tablinks active" onClick="openTab(event,'setup')">Setup</button>`
The tab button points to the `content_div` with id `setup` on click and the id for the button itself is `content_id(i.e setup)+"Link"` which becomes `setupLink`