Use `_pbench_` prefixed env var for host names
Instead of accepting `hostname` and `full_hostname` as pre-defined
environment variables, we rename them to be `_pbench_hostname` (contains
`$(hostname -s)`) and `_pbench_full_hostname` (contains
`$(hostname -f)`).  This prevents an inadvertent environment variables
of the same name from change the behavior of the agent.

We also consistently use those environment variables where possible to
avoid disagreements on the host name used.

Further we remove a `FIXME` comment in `agent/base` which was addressed
in PR #1248.

Some of the Python 3 code changed, and so `flake8` detected a problem.
Why it wasn't detected before is a mystery, but we fix it here as well.