Fix agent side `test_move_results`
While working on PR #1948, refactoring the common logging infrastructure,
the tests in `test_move_results` were failing for odd reasons.  Looking
at them a bit more closely it became apparent the test was not actually
invoking the API's PUT method.  This commit creates a test that allows
the API to be invoked, fixing the move results code to work correctly.