Rework pass-thru API implementation a bit
This all started because logging was not quite set up correctly for the
server side API tests.  We use the `caplog` fixture to figure out what
is being logged and then added `file` based logging by default so that
logs would not leak through to the local system logs via the default
`devlog` server configuration during a test run.

In getting the tests to work, we adjust how exceptions are handled so
that emitted messages are not so broad.  We use 5xx status codes
properly and quiet the logging where possible.

When an internal error is hit, we only report "INTERNAL ERROR" in the
response, but log the details.