Restore the SELinux context of the tarball after the move
Commit c3ad67c101e394784924f14c5bfceb8d239b636f changed the way
we get the tarball from the reception area to the archive, from `cp`
to `mv` in order to avoid a data copy. But `cp' set the SELinux context
from the parent directory, which made the tarball accessible through HTTP;
`mv` does not do that.

Add an explicit `restorecon` to remedy that and fix test-5.2 for the
new output.