Remove colorlog; add Makefile for pbench-devel
The Python 3 `colorlog` module was removed with commit 3a851f84, but
we failed to remove it from the `pbench-agent` RPM requires list, or
from the recently added `pbench-devel` container image.

While we are at it, we added a simple Makefile for building the
`pbench-devel` container image and pushing it to Quay.io.

For this `b0.70` branch we also have to remove the use of `colorlog`
from `setup_logging` in `lib/pbench/agent/utils.py`.