Remove colorlog; add Makefile for pbench-devel
The Python 3 `colorlog` module was removed with commit 3a851f84, but
we failed to remove it from the `pbench-agent` RPM requires list, or
from the recently added `pbench-devel` container image.

While we are at it, we added a simple Makefile for building the
`pbench-devel` container image and pushing it to Quay.io.