Enhance `tool-scripts/README`, `doc/CONTRIBUTING`
We needed a way to document how to install the new `node-exporter` and
`dcgm` tools now available as tools to be registered.  But the details
of such installation steps don't belong in our official documentation.

Instead, we add a significantly enhanced `agent/tool-scripts/README.md`
file which explains the nature of pbench tools in more detail, providing
a context to provide the esoteric installation procedures for these
tools.

While we are at it, we also:

 * Collapse the two "contributing" guides by moving the content of
   the original `doc/CONTRIBUTING.md` and moving it into the top-level
   `README.md`, and renaming the `doc/CONTRIBUTION.md` to be the more
   active call-to-action `CONTRIBUTING.md`

 * Fix the Prometheus GitHub reference to in the place-holder
   `node-exporter` tool script

 * Fix the URL reference to the Pbench Installation Guide in the
   `doc/INSTALL.md` file