Harden against spaces in file names
When using the `bash` `for <var> in $(ls -1d ...); do` expressions, the
shell will assign the variable values not only split by new-lines, but
also by spaces.  So if a file name or directory in the path has a space
in the name, the `ls -1d` output will not be interpreted correctly.