More restorecon-related fixes
This is yet another continuation to PR #1973 and #1960.

This fixes the tar ball and MD5 paths (which *were* caught by
unit test 5.2, except that I was asleep at the wheel).

It also adds /usr/sbin to the PATH: when the script is executed
under cron, that is not (always?) part of the inherited PATH
and the script complains that restorecon cannot be found.