Fix spec file and ansible roles
When I simplified the spec file back in June, I went a bit too far
and lost the js files that we need for visualization. Some of these have
been added back: the rest are needed for the (very) old web-server version
only.

The firewall role added http and https services to the "permanent" setup
but neglected to reload the rules, so they were not valid for the current
boot (although they would work fine after a reboot :-) ).

The pbench-tarballs link from /var/www/html to the archive was not created
by the playbook (it was added manually to the production server). It is
now added by the playbook.