Add `--controller` option to `pbench-move-results`
This addresses the sentiment in issue #1945.

Adding a new switch, `--controller` to the `pbench-move-results` allows
a user to set the controller name when pushed to the remote server,
while preserving the controller name used originally.  We also take the
opportunity to record the original short and long hostname values
explicitly, and the host's IP address (output of `hostname -i`).

We also add support for the `PBENCH_CONTROLLER` environment variable
for those that don't want to use the switch (similar to `--user`).

We also:

 * Fix `pbench-move-results` to take no actions remotely if there are
   no result directories to be moved
 * Fix the tabs vs spaces formatting in the three files related to
   `pbench-move-results`
 * Add a number of unit tests back that were left dangling, writing
   a check to make sure this doesn't happen in the future
 * Add additional unit tests to verify the new behaviors