Fix SELinux labels on the pbench directory structure
We need to use `semanage' to set the correct labels on the top level
pbench directory and its children: most of them get `default_t' but
`archive' and `public_html' and their children get `httpd_sys_content_t'.

The sefcontext ansible module is needed for this: it's available from
Galaxy:

    ansible-galaxy collection install community.general

The rest of the tasks create the directory structure (I deleted the
setype: setting since it does not survive a restorecon - I now assume
that it is just doing the equivalent of a chcon). At the end we run a
recursive restorecon to set the labels on everything.