Fix the Python 3 logger to handle time properly
While working PR #1914, we found that our log formatter was not handling
records that already had a formatted `asctime` field.  When we compared
the code against what the default Python 3 (3.6 version used, the lowest
version we support), we found that we were not handling this properly.

This change updates the code to more closely match the behavior of the
3.6 logger using the logging interfaces as intended.  Well, with one
exception, we ignore the specified `datafmt` and always use UTC time in
the ISO format.