Refactor Tool Meister infra for data capture
The last set of `ssh` commands issued by the pbench-agent that are not
involved in orchestrating the Tool Meisters, outside of those issued by
the `bench-scripts`, are for `hostname -s` data collection on all
the registered remote tool hosts.

The Tool Meisters are now given the responsibility for collecting
information about their environment, including the version of the
pbench agent, and a more comprehensive collection of host name
information, returning that with their startup acknowledgement
payload.  This change also addresses a previous bug where registered
host labels were not being used in the on-disk data capture
directories as they were before.

This collected data set is now recorded by the Tool Data Sink, after
restructuring the action flow so that clients send action payloads to
the Tool Data Sink, which then forwards those payloads to the Tool
Meisters.  This gives the Tool Data Sink the ability to collect all the
Tool Meister information and write it to disk before acknowledging to
the client that startup is complete.

We add an entirely new section in the `metadata.log` file which is
per-host/per-tool, recording the tool options, and the output from the
tool install checks (if any).  This section supercedes the previous
tool options in the per-host sections, though we leave those options
there for compatibility for now.

We add graceful handling of bad PUT requests from Tool Meisters to the
Tool Data Sink.

With this refactoring, we also collapse all the "sysinfo", "init", and
"end" actions into the respective `pbench-tool-meister-start` and
`pbench-tool-meister-stop` interfaces to simplify the CLI behaviors.

To make this possible, we add a new `Client` API used by the CLI
interfaces `pbench-tool-meister-start` and `pbench-tool-meister-stop`.

We have also undertaken a major refactoring of `pbench-sysinfo-dump` to
remove the dependency on `base` for environment variables.  As a result
of this work, we drop the stockpile configuration entirely, as it is no
longer required.

The Tool Meister unit tests were enhanced to use persistent tools, and
the unit test framework now properly cleans up the Tool Meister test
environment when tests fail.

We moved the pbench-sysinfo-dump CLI command to the
`agent/util-scripts/tool-meister` directory so that users won't see
that internal command in their `PATH`.

We updateed the UML Seq Diag to reflect the code, addng the
`init`/`end`/`sysinfo` actions, and the fact that the Tool Data Sink
is now the gatekeeper of actions, forwarding them on to the Tool
Meisters instead of the client sending actions directly to the Tool
Meisters.

Since we no longer have the CLI command `pbench-collect-sysinfo`, we
drop it from the unit tests, and replace it with more appropriate tests.

To that end:
  * tests 25 - 30 now just invoke `pbench-verify-sysinfo-options`
  * tests 54 & 55 just invoke the `--help` option on
    `pbench-tool-meister-start` and `-stop`
  * tests 23 * 24 are dropped

The Tool Meisters need to start their persistent tools first before the
collectors can run correctly.  The PCP pmlogger collectors wait for the
remote pmcds to start listening on the expected ports.  We have to pass
along the "init" action to the Tool Meisters before we setup the PCP
collectors.

The `pbench-tool-meister-stop` command will now return an error when the
`end` operation fails, or if we can't create the directory for the `end`
operation to work.  We always wait for the local Tool Data Sink and
local Tool Meister to exit before killing the Redis server, regardless
of the success or failure of `terminate` operation.