Fix py move-results to use `ConfigParser`
The Python 3 based "move-results" code was using the old `configtools`
module instead of just using `ConfigParser` directly.  This would end
up causing failures in the unit tests because it was calling the
`parse_args` method of `configtools` to first parse the command line
arguments.  But in principle, we never want to access the command line
arguments while running unit tests.  The unit test should provide them
for the method under test to properly ensure the correct behavior is
tested.

We simply replace the use of `configtools` with `ConfigParser` directly,
which ends up simplifying the code as well.

Further, using configtools just for the result directory `metadata.log`
is over kill, because `configtools` was there to support multiple
configuration files, and there is only one `metadata.log` file.

We remove the `SETUP` and `TEARDOWN` messages since they are really
debugging statements and have no business being there for normal test
operation.