Fix bug when reading a file in chunks
Fix #2034

Commit 8a4417838959ec17d234eff887540dc461386cb2 broke the function
calculate_multipart_etag(): it replaced a (working) while-loop that
was reading the tarball in chunks with a broken for-loop which read
one chunk of the file and then iterated over the bytes of that chunk.

We retain the for-loop but fix the iterator to give us back the next
chunk each time.