Fix issue with stranded pidstat and turbostat
For `turbostat` we use redirection to sub-shells instead of pipes so
that we can exec the datalog process allowing it to be killed directly.

This was not a problem when we were using `screen` because it would
clean up all the sub-processes for us.  With the Tool Meister we no
longer use `screen`.

In the case of the `pidstat` and `turbostat` tools, we were not
directly killing the process generating data. This allowed
the other processes in the pipeline to hang around unexpectedly.

However, cannot use the same method for `pidstat` employed by the
`turbostat` fix.  The use of sub-shell redirection in the `pidstat`
case doesn't let the operation of the `pidstat-convert` complete by
the time `pidstat` exits.  Instead, we re-write `pidstat-convert`
to wrap `pidstat` instead.

Fixes issue #2026.