When installing the server, set the host-info state to maintenance
Currently, when installing the server, we set the host-info state to active,
which allows pbench-move/copy-results to start sending tar balls to the server.

This PR sets the host-info state to maintenance.

It also adds playbooks and roles to explicitly enable and disable the
server. The workflow now becomes

   ansible-playbook ... pbench-server-disable.yml

   ansible-playbook ... pbench-server-install.yml
   # check crontab
   ansible-playbook ... pbench-server-enable.yml

Unfortunately, the disable role has no easy way to make sure that the
queue is drained: it can check the reception area to make sure it's
empty, but various cron jobs may still be running processing tar balls
(or might run in the next few minutes): it could check all the state
directories to make sure they don't contain any tar ball symlinks, but
that seems a bit fragile (and perhaps expensive). So, for now at
least, we punt and leave it up to the administrator to check and make
sure that, after the disablement, all the tar balls have been
processed before shutting off the crontab manually with

   crontab -u pbench -r

Another possibility is to shut off the crontab anyway, and wait until
all currently running cron jobs are done before proceeding. There
might still be some tar balls in a to-process state, but they will be
processed after the server is enabled again.

This is what's implemented here: enabling the server turns the crontab
on and disabling the server turns the crontab off. The administrator
still has to wait, checking that the various pbench processes that
might be running have actually finished, before proceeding.

The crontab enablement/disablement is done by separate roles: it's
only the playbooks that impose the unitarity of enablement or
disablement. After review, the roles are now protected, so that they
only trigger when the roles are run as the "pbench" user: they are no-ops
otherwise.

Privilege escalation in all the playbooks is now done through the
remote_user variable. It is `root` for the installation playbook,
but it is `pbench` for the enablement and disablement playbooks.

Also save the crontab before disabling.