Default configuration for Apache reverse-proxy from port 80
The Pbench server API listens on port 8001; we want to be able to access
the API externally through normal HTTP/HTTPS.

With an external reverse proxy gateway (e.g., NGINX) this can be accomplished
by redirecting (e.g.) pbench.example.com:/api/ to the actual server at
port 8001 and opening the 8001/tcp port in the server firewall.

Our default Ansible deployment mechanism configures a local Apache which
is directly accessed from outside. This PR adds an Apache reverse proxy
to allow external API access on port 80.

With these changes, the Pbench gunicorn wsgi server will still listen on
port 8001, but this port won't need to be opened externally as Apache
will proxy /api/ to port 8001. This means that all external accesses will
go to port 80 (http) or (when we get there) 433 (https).

In the short term this affects the pbench_server configuration in the
dashboard endpoints.js. Once #2024 (server-side endpoint configuration
and metadata) is merged, I plan to change the dashboard to use that,
simply fetch-ing that from window.origin and assigning it to window.endpoints
within the launch html page, removing the need to manage a local static
endpoints.js entirely.